+++ examples/for-tests-react-16/src/App.js
@@ -222,6 +222,7 @@ let recipeList = [
 
 const testContext = {
     disableDefaultUI: getQueryParams("disableDefaultUI") === "true",
+    thirdPartyRedirectURL: localStorage.getItem("tp-redirectURL"),
 };
 
 if (authRecipe === "thirdparty") {
@@ -582,7 +583,7 @@ function getEmailPasswordConfigs({ disableDefaultUI }) {
     });
 }
 
-function getThirdPartyPasswordlessConfigs({ disableDefaultUI }) {
+function getThirdPartyPasswordlessConfigs({ disableDefaultUI, thirdPartyRedirectURL }) {
     return ThirdPartyPasswordless.init({
         style: {
             container: {
@@ -721,6 +722,7 @@ function getThirdPartyPasswordlessConfigs({ disableDefaultUI }) {
                 {
                     id: "auth0",
                     name: "Auth0",
+                    getRedirectURL: thirdPartyRedirectURL !== null ? () => thirdPartyRedirectURL : undefined,
                 },
             ],
             defaultCountry: passwordlessDefaultCountry,
@@ -820,7 +822,7 @@ function getPasswordlessConfigs({ disableDefaultUI }) {
     });
 }
 
-function getThirdPartyConfigs({ disableDefaultUI }) {
+function getThirdPartyConfigs({ disableDefaultUI, thirdPartyRedirectURL }) {
     return ThirdParty.init({
         style: {
             container: {
@@ -914,6 +916,7 @@ function getThirdPartyConfigs({ disableDefaultUI }) {
                 {
                     id: "auth0",
                     name: "Auth0",
+                    getRedirectURL: thirdPartyRedirectURL !== null ? () => thirdPartyRedirectURL : undefined,
                 },
             ],
         },
@@ -924,7 +927,7 @@ function getThirdPartyConfigs({ disableDefaultUI }) {
     });
 }
 
-function getThirdPartyEmailPasswordConfigs({ disableDefaultUI }) {
+function getThirdPartyEmailPasswordConfigs({ disableDefaultUI, thirdPartyRedirectURL }) {
     return ThirdPartyEmailPassword.init({
         preAPIHook: async (context) => {
             console.log(`ST_LOGS THIRD_PARTY_EMAIL_PASSWORD PRE_API_HOOKS ${context.action}`);
@@ -1110,6 +1113,7 @@ function getThirdPartyEmailPasswordConfigs({ disableDefaultUI }) {
                 {
                     id: "auth0",
                     name: "Auth0",
+                    getRedirectURL: thirdPartyRedirectURL !== null ? () => thirdPartyRedirectURL : undefined,
                 },
             ],
         },