+++ examples/for-tests/src/App.js
@@ -232,6 +232,7 @@ let recipeList = [
 
 const testContext = {
     disableDefaultUI: getQueryParams("disableDefaultUI") === "true",
+    thirdPartyRedirectURL: localStorage.getItem("thirdPartyRedirectURL"),
 };
 
 if (authRecipe === "thirdparty") {
@@ -599,7 +600,7 @@ function getEmailPasswordConfigs({ disableDefaultUI }) {
     });
 }
 
-function getThirdPartyPasswordlessConfigs({ disableDefaultUI }) {
+function getThirdPartyPasswordlessConfigs({ disableDefaultUI, thirdPartyRedirectURL }) {
     return ThirdPartyPasswordless.init({
         style: {
             container: {
@@ -750,6 +751,7 @@ function getThirdPartyPasswordlessConfigs({ disableDefaultUI }) {
                 {
                     id: "auth0",
                     name: "Auth0",
+                    getRedirectURL: thirdPartyRedirectURL !== null ? () => thirdPartyRedirectURL : undefined,
                 },
             ],
             defaultCountry: passwordlessDefaultCountry,
@@ -858,7 +860,7 @@ function getPasswordlessConfigs({ disableDefaultUI }) {
     });
 }
 
-function getThirdPartyConfigs({ disableDefaultUI }) {
+function getThirdPartyConfigs({ disableDefaultUI, thirdPartyRedirectURL }) {
     return ThirdParty.init({
         style: {
             container: {
@@ -965,6 +967,7 @@ function getThirdPartyConfigs({ disableDefaultUI }) {
                 {
                     id: "auth0",
                     name: "Auth0",
+                    getRedirectURL: thirdPartyRedirectURL !== null ? () => thirdPartyRedirectURL : undefined,
                 },
             ],
         },
@@ -975,7 +978,7 @@ function getThirdPartyConfigs({ disableDefaultUI }) {
     });
 }
 
-function getThirdPartyEmailPasswordConfigs({ disableDefaultUI }) {
+function getThirdPartyEmailPasswordConfigs({ disableDefaultUI, thirdPartyRedirectURL }) {
     return ThirdPartyEmailPassword.init({
         preAPIHook: async (context) => {
             if (localStorage.getItem(`SHOW_GENERAL_ERROR`) === `THIRD_PARTY_EMAIL_PASSWORD ${context.action}`) {
@@ -1174,6 +1177,7 @@ function getThirdPartyEmailPasswordConfigs({ disableDefaultUI }) {
                 {
                     id: "auth0",
                     name: "Auth0",
+                    getRedirectURL: thirdPartyRedirectURL !== null ? () => thirdPartyRedirectURL : undefined,
                 },
             ],
         },